requirements.txt
click==8.0.0
Flask==2.0.0
Flask-MySQLdb==0.2.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
mysqlclient==2.0.3
python-dotenv==0.17.1
Werkzeug==2.0.0
